Update owncloud.py
I tried this on my Raspberry Pi running Xbian and my ownCloud 6.0.2 and I was getting an error like "Cannot login, incorrect password..." (or something like this) and after debugging I was able to fix it by adding "index.php" to line 96. 

Then when i was able to connect it listed all my folders but with the urlencoding. First, they didn't look good. Second, those folders won't open, since they the original encoding gets encoded again creating incorrect folders. ANd the solution is to remove the encoding using "urllib.unquote" (as in line 211 of the modified file) before displaying the folders.